Rename lineal-viz to @lineal-viz/lineal
It's not the org name I wanted, but what can you do?
