+++ package.json
@@ -5,22 +5,22 @@
   "license": "MIT",
   "author": "",
   "scripts": {
-    "prepare": "yarn workspace lineal-viz run prepare",
+    "prepare": "yarn workspace @lineal-viz/lineal run prepare",
     "start": "npm-run-all --parallel start:*",
-    "start:addon": "yarn workspace lineal-viz run start",
+    "start:addon": "yarn workspace @lineal-viz/lineal run start",
     "start:test-app": "yarn workspace test-app run start",
     "start:docs": "yarn workspace docs run start",
     "lint": "npm-run-all --aggregate-output --continue-on-error --parallel \"lint:!(fix)\"",
-    "lint:addon": "yarn workspace lineal-viz run lint",
+    "lint:addon": "yarn workspace @lineal-viz/lineal run lint",
     "lint:test-app": "yarn workspace test-app run lint",
     "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel lint:fix:*",
-    "lint:fix:addon": "yarn workspace lineal-viz run lint:fix",
+    "lint:fix:addon": "yarn workspace @lineal-viz/lineal run lint:fix",
     "lint:fix:test-app": "yarn workspace test-app run lint:fix",
     "test": "npm-run-all --aggregate-output --continue-on-error --parallel \"test:!(watch)\"",
     "test:watch": "npm-run-all --aggregate-output --continue-on-error --parallel test:watch:*",
     "test:test-app": "yarn workspace test-app run test",
     "test:watch:test-app": "yarn workspace test-app run test:watch",
-    "test:watch:addon": "yarn workspace lineal-viz run start"
+    "test:watch:addon": "yarn workspace @lineal-viz/lineal run start"
   },
   "devDependencies": {
     "npm-run-all": "^4.1.5",
